requirements.txt
yoyo-migrations==8.2.*
python-dotenv==1.0.*
PyMySQL==1.1.*
cryptography==41.0.*
